+++ lib/src/mcu/palettes/tonal_palette.dart
@@ -16,6 +16,7 @@ import 'package:collection/collection.dart' show ListEquality;
 import 'package:meta/meta.dart';
 
 import '../hct/hct.dart';
+import 'key_color.dart';
 
 /// A convenience class for retrieving colors that are constant in hue and
 /// chroma, but vary in tone.
@@ -74,13 +75,13 @@ class TonalPalette {
   // ignore: sort_constructors_first, we prefer this order.
   TonalPalette._fromHueAndChroma(this.hue, this.chroma)
       : _cache = <int, int>{},
-        keyColor = createKeyColor(hue, chroma),
+        keyColor = KeyColor(hue, chroma).create(),
         _isFromCache = false;
 
   // ignore: sort_constructors_first, we prefer this order.
   TonalPalette._fromCache(Map<int, int> cache, this.hue, this.chroma)
       : _cache = cache,
-        keyColor = createKeyColor(hue, chroma),
+        keyColor = KeyColor(hue, chroma).create(),
         _isFromCache = true;
 
   /// Create colors using [hue] and [chroma].
@@ -122,46 +123,6 @@ class TonalPalette {
       }
     }
     return TonalPalette._fromCache(cache, bestHue, bestChroma);
-  }
-
-  /// Creates a key color from a [hue] and a [chroma].
-  /// The key color is the first tone, starting from T50, matching the
-  /// given hue and chroma. Key color [Hct].
-  static Hct createKeyColor(double hue, double chroma) {
-    const double startTone = 50.0;
-    Hct smallestDeltaHct = Hct.from(hue, chroma, startTone);
-    double smallestDelta = (smallestDeltaHct.chroma - chroma).abs();
-    // Starting from T50, check T+/-delta to see if they match the requested
-    // chroma.
-    //
-    // Starts from T50 because T50 has the most chroma available, on
-    // average. Thus it is most likely to have a direct answer and minimize
-    // iteration.
-    for (double delta = 1.0; delta < 50.0; delta += 1.0) {
-      // Termination condition rounding instead of minimizing delta to avoid
-      // case where requested chroma is 16.51, and the closest chroma is 16.49.
-      // Error is minimized, but when rounded and displayed, requested chroma
-      // is 17, key color's chroma is 16.
-      if (chroma.round() == smallestDeltaHct.chroma.round()) {
-        return smallestDeltaHct;
-      }
-
-      final Hct hctAdd = Hct.from(hue, chroma, startTone + delta);
-      final double hctAddDelta = (hctAdd.chroma - chroma).abs();
-      if (hctAddDelta < smallestDelta) {
-        smallestDelta = hctAddDelta;
-        smallestDeltaHct = hctAdd;
-      }
-
-      final Hct hctSubtract = Hct.from(hue, chroma, startTone - delta);
-      final double hctSubtractDelta = (hctSubtract.chroma - chroma).abs();
-      if (hctSubtractDelta < smallestDelta) {
-        smallestDelta = hctSubtractDelta;
-        smallestDeltaHct = hctSubtract;
-      }
-    }
-
-    return smallestDeltaHct;
   }
 
   /// Returns a fixed-size list of ARGB color ints for common tone values.